routers/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic[email]
passlib
bcrypt==4.0.1
python-jose[cryptography]